// src/app/interfaces/user.ts
export class User {
  public idIglesia: number;
  public usuario: string;
  public pass: string;
  idTipoServicio?: number;
  topic?: string;
  idUserType: number;
  isSuperUser: boolean;
  iglesia: string;
  idUsuario: number;
  nombre: string;
  apellido: string;
  picture: string;
  logoPic: string;
  idUserOrganization?: number;
  is_valid?: boolean;
  service_type_picture?: string;
  email: string;
}